requirements.txt
PyYAML

crossbar
autobahn

sanic

aioredis
requests

discord.py
